asns/asn3/adminPCQt/pyproject.toml
[project]
name = "adminpcqt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic>=2.12.5",
    "pyside6>=6.10.1",
]
